Wrapped Features to make it possible to switch out etcd
Moved ACL into Domain
