+++ server/app.js
@@ -11,7 +11,7 @@ var auditRoutes = require('./routes/auditRoutes');
 var applicationRoutes = require('./routes/applicationRoutes');
 var featureRoutes = require('./routes/featureRoutes');
 var path = require('path');
-var acl = require('./acl');
+var acl = require('./domain/acl');
 var config = require('./../config/config.json');
 var _ = require('underscore');
 var passport = require('./auth').init(config);
