+++ server/routes/authorisationRoutes.js
@@ -2,7 +2,7 @@
 
 var etcd = require('../etcd');
 var _ = require('underscore');
-var acl = require('../acl');
+var acl = require('../domain/acl');
 var validator = require('validator');
 var config = require('./../../config/config.json');
